bl size_t
git-svn-id: svn+ssh://astrometry.net/svn/trunk/src/astrometry@24963 03a3cea6-2c03-0410-ac9b-9271b1c66c29

+++ util/an-opts.c
@@ -58,7 +58,7 @@ void opts_print_help(bl* opts, FILE* fid,
 }
 
 int opts_getopt(bl* opts, int argc, char** argv) {
-    int i, j, N;
+    size_t i, j, N;
     char* optstring;
     int c;
     struct option* longoptions;
